+++ src/main/java/com/ham/netnovel/comment/Comment.java
@@ -2,12 +2,11 @@ package com.ham.netnovel.comment;
 
 
 import com.ham.netnovel.commentLike.CommentLike;
-import com.ham.netnovel.commentLike.LikeType;
+import com.ham.netnovel.commentLike.data.LikeType;
 import com.ham.netnovel.episode.Episode;
 import com.ham.netnovel.member.Member;
 import com.ham.netnovel.reComment.ReComment;
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
