+++ src/main/java/com/ham/netnovel/commentLike/data/LikeType.java
@@ -1,4 +1,4 @@
-package com.ham.netnovel.commentLike;
+package com.ham.netnovel.commentLike.data;
 
 public enum LikeType {//좋아요 타입
     LIKE,//좋아요
