+++ src/main/java/com/ham/netnovel/reComment/ReCommentLike.java
@@ -1,6 +1,6 @@
 package com.ham.netnovel.reComment;
 
-import com.ham.netnovel.commentLike.LikeType;
+import com.ham.netnovel.commentLike.data.LikeType;
 import com.ham.netnovel.member.Member;
 import com.ham.netnovel.reCommentLike.ReCommentLikeId;
 import jakarta.persistence.*;
