+++ src/main/java/com/ham/netnovel/reCommentLike/ReCommentLikeToggleDto.java
@@ -1,7 +1,7 @@
 package com.ham.netnovel.reCommentLike;
 
 
-import com.ham.netnovel.commentLike.LikeType;
+import com.ham.netnovel.commentLike.data.LikeType;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
